project_akka/requirements.txt
# Project Akka Dependencies
# Python 3.12 - NVIDIA Jetson Orin Nano

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Validation
pydantic>=2.5.0

# LLM Services
google-generativeai>=0.3.0

# Configuration
pyyaml>=6.0.1

# HTTP Client (for Ollama)
httpx>=0.25.0

# Ollama
ollama>=0.6.1

# Optional: Development tools
# python-dotenv>=1.0.0
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
